requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart>=0.0.9

pydantic>=2.8
python-dotenv>=1.0

# AutoGen 0.6+ (AgentChat + Azure/OpenAI clients)
autogen-agentchat>=0.6.2
autogen-ext[openai,azure]>=0.6.2

# MCP server + client (Python SDK)
mcp[cli]>=1.2.0
httpx>=0.27

# Parsing
pypdf>=5.0.0
python-docx>=1.1.2
beautifulsoup4>=4.12

# Simple retrieval (BM25)
rank-bm25>=0.2.2
nltk>=3.9

# Frontend
streamlit>=1.36

# Optional observability
langsmith>=0.1.83
